simulation/requirements.txt
kafka-python>=2.0
pydantic>=2.0
python-dotenv>=1.0
numpy>=1.24
Pillow>=10.0
